+++ src/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  HTML: require('./html'),
+  Text: require('./text'),
+  Editor: require('./editor'),
+  Container: require('./container')
+};